requirements.txt

python-telegram-bot>=20.0
flask>=2.0
python-dotenv>=1.0
google-generativeai>=0.3.0
PyPDF2>=3.0
python-docx>=0.8.11
